backend/requirements.txt
absl-py==1.0.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==4.1
django-cors-headers==4.1.0
djangorestframework==3.13.0
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio
h5py
idna==2.8
joblib==0.14.1
Keras==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.22
oauthlib==3.1.0
opencv-python==4.7.0.72
opt-einsum==3.1.0
pandas==1.2.4
Pillow
protobuf
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn
scipy
six==1.13.0
sqlparse==0.3.0
tensorboard==2.12.0
tensorflow==2.12
tensorflow-estimator==2.12.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==1.0.1
wrapt==1.11.2